+++ lib/rollout/notifications/notifiers/status_change.rb
@@ -10,14 +10,17 @@ class Rollout
 
         def notify(feature_name, new_status, new_percentage=nil)
           @channels.each do |c|
-            publish_for_slack_channel(c, feature_name, new_status, new_percentage) if c.type == :slack
-            publish_for_email_channel(c, feature_name, new_status, new_percentage) if c.type == :email
+            if c.class == Rollout::Notifications::Channels::Email
+              publish_to_email_channel(c, feature_name, new_status, new_percentage)
+            else
+              publish_to_channel(c, feature_name, new_status, new_percentage)
+            end
           end
         end
 
         private
 
-        def publish_for_slack_channel(c, feature_name, new_status, new_percentage)
+        def publish_to_channel(c, feature_name, new_status, new_percentage)
           if new_status == :activated
             text = "Feature flag '#{feature_name}' has been activated with percentage #{new_percentage}!"
           elsif new_status == :deactivated
@@ -26,7 +29,7 @@ class Rollout
           c.publish(text)
         end
 
-        def publish_for_email_channel(c, feature_name, new_status, new_percentage)
+        def publish_to_email_channel(c, feature_name, new_status, new_percentage)
           if new_status == :activated
             subject = 'Feature flag has been activated!'
             content = "Feature flag '#{feature_name}' has been activated with percentage #{new_percentage}!"